GH-81 Move yaml converter function from metadata_io.hpp to cpp file
